+++ src/routes/(v2)/v2/app-error/schema.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+
+export const userSchema = z.object({
+	name: z.string(),
+	exception: z.boolean()
+});